+++ src/ergotest/test_runner_worker_process.ts
@@ -1,6 +1,6 @@
 // Copyright Titanium I.T. LLC. License granted under terms of "The MIT License."
 import { importRendererAsync, TestSuite } from "./test_suite.js";
-import { TestCaseResult, TestMark, TestResult, TestSuiteResult } from "./test_result.js";
+import { RunResult, TestCaseResult, TestMark, TestResult, TestSuiteResult } from "./test_result.js";
 import { Clock } from "../infrastructure/clock.js";
 import process from "node:process";
 import { WorkerInput } from "./test_runner.js";
@@ -38,11 +38,10 @@ async function runWorkerAsync(
 	try {
 		const renderError = await importRendererAsync(renderer);
 
-		process.on("uncaughtException", (err) => {
+		process.on("uncaughtException", (error) => {
 			const errorResult = TestResult.suite([], [
-				TestResult.fail("Unhandled error in tests", err, {
-					mark: TestMark.none,
-					renderError,
+				TestResult.testCase({
+					it: RunResult.fail({ name: "Unhandled error in tests", error, renderError }),
 				}),
 			]);
 			sendFinalResult(errorResult, cancelKeepAliveFn);